# ioq benchmarking and performance improvements