--- a/src/ioq.h
+++ b/src/ioq.h
@@ -8,6 +8,7 @@
 #ifndef BFS_IOQ_H
 #define BFS_IOQ_H
 
+#include "bfs.h"
 #include "dir.h"
 #include "stat.h"
 
@@ -45,18 +46,11 @@ enum ioq_nop_type {
 };
 
 /**
- * The I/O queue implementation needs two tag bits in each pointer to a struct
- * ioq_ent, so we need to ensure at least 4-byte alignment.  The natural
- * alignment is enough on most architectures, but not m68k, so over-align it.
- */
-#define IOQ_ENT_ALIGN alignas(4)
-
-/**
  * An I/O queue entry.
  */
 struct ioq_ent {
 	/** The I/O operation. */
-	IOQ_ENT_ALIGN enum ioq_op op;
+	cache_align enum ioq_op op;
 
 	/** The return value (on success) or negative error code (on failure). */
 	int result;